requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
requests==2.28.1
urllib3==1.26.13
